refactor: extract __main__.py components into dedicated modules
    Non-functional reorganization: move cohesive groups of code out of the
    __main__.py into dedicated modules. No behavior change.

    New/extended modules:
    - stats.py: status/emoji maps, Stat, collect_stat,
      calculate_summary_stats, print_stats, load_stats, save_stats
    - runners/harness.py: TestRunner ABC, PytestRunner, HammettRunner,
      ListAllTestsResult, collected_test_names, unused, and the
      test-runner exceptions (Collect/BadTestExecutionCommands)
    - ui/browse.py: ResultBrowser Textual app wrapped in
      run_result_browser(); get_diff_for_mutant/apply_mutant are injected
      to avoid a circular import back into __main__. Uses upstream #543's
      @work/Lock/get_current_worker diff-loading model.
    - ui/terminal.py: spinner + status_printer/print_status
    - utils/file_utils.py: walk_all_files/walk_source_files/
      walk_mutatable_files, copy_src_dir, copy_also_copy_files,
      setup_source_paths

    Move result_browser_layout.tcss into ui/ so Textual's CSS_PATH
    resolves relative to ui/browse.py.

    __main__.py re-imports the moved public names, so external
    `from mutmut.__main__ import ...` call sites keep working unchanged.
    Repoint code_coverage.py's TYPE_CHECKING import to runners.harness.
