+++ src/mutmut/ui/__init__.py
@@ -1,0 +1,1 @@
+"""UI components for mutmut."""
